Moves game_info to be a hash in hangman.rb
Before it was in game_info.rb, but there were no other methods. It was
just being stored in another file, with the idea that I would move some
methods over there later. But now that I want to save game files, I
figure that saving the game_info object will be easier if it's just
stored as a hash.